Created new trait for select scopes
