--- .styleci.yml
@@ -1,3 +1,0 @@
-preset: psr2
-
-linting: true
